refactor(identity): which account a verified tuple belongs to lives beside, not inside, the handlers
